Non-class SerdeTypes in lookup; expand roots; strip backslashes
This means references to e.g. primitives are unique (so the total lookup size reduces). Currently, circular references are only broken
at SerdeTypeClass (which adds a partial value into the lookup and amends it afterwards), so in theory it's possible to have duplicates
of other nested types (e.g. unions, including nullable types, and arrays.

Expanded root types means that if a root type is complex (e.g. a union, an array, or a discriminated type) then the inner types /
subtypes are also considered root types (so warnings aren't displayed when using them).