Empty array serde: don't ignore, use {} for assoc
Previously, the code liberally used empty arrays as default values for array-type properties, then ignored them when normalizing.
This was for two main reasons:
a) It kept the JSON sent to Wiremock simple, and clear of values that have no effect
b) It avoided the mismatch between JSON empty arrays [] and objects {}, and PHP arrays and associative arrays being the same thing
  (only disambiguated by values, so empty arrays / assoc arrays are identical, so json_encode to "[]" even though you may need "{}")

This change stops ignoring empty array values when normalizing (see the SerdeTypeClass change), which means two things:
a) many properties need to be typed as nullable arrays, defaulted to null. (Note that $arr[] = 'foo' on when $arr is null automatically creates
  and array, in PHP). See the various changes to classes in e.g. WireMock\Client.
b) normalizing needs to provide an empty stdClass when normalizing an assoc array typed empty array. See the change in Serializer